+++ .vs/rufus.vcxproj
@@ -405,6 +405,7 @@
     <ClInclude Include="..\src\dev.h" />
     <ClInclude Include="..\src\ui.h" />
     <ClInclude Include="..\src\ui_data.h" />
+    <ClInclude Include="..\src\vhd.h" />
     <ClInclude Include="..\src\winio.h" />
     <ClInclude Include="..\src\wue.h" />
   </ItemGroup>
