+++ .vs/rufus.vcxproj.filters
@@ -188,6 +188,9 @@
     <ClInclude Include="..\src\wue.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\src\vhd.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="..\res\rufus.ico">
